Basic housekeeping (attach database file to LocalDb instance, refactor solution name, refactor project name, refactor namespaces; update assembly properties; enforce one-class-per-file rule).
