+++ Refaction.Service/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Refaction.Service.WebApiApplication" Language="C#" %>
