+++ Refaction.Service/Helpers.cs
@@ -1,7 +1,7 @@
 ﻿using System.Data.SqlClient;
 using System.Web;
 
-namespace refactor_me.Models
+namespace Refaction.Service
 {
     public class Helpers
     {
