--- refactor-me/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="refactor_me.WebApiApplication" Language="C#" %>
